requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
anthropic>=0.25.0
 # Optional (alternate design): LangChain + LangGraph
 langchain>=0.0.300 ; extra == "langchain"
 langgraph ; extra == "langgraph"